+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 1.0.0+dev
 - added exception message in method SimpleContainer::get
+- improved exception messages
 
 Version 1.0.0
 - initial version